+++ src/Newman.js
@@ -6,6 +6,7 @@ var jsface          = require("jsface"),
     Globals          = require('./utilities/Globals'),
     Options          = require('./utilities/Options'),
     log              = require('./utilities/Logger'),
+    importer         = require('./utilities/Importer'),
     fs               = require('fs'),
     exec             = require('child_process').exec;
 
@@ -25,6 +26,9 @@ var Newman = jsface.Class([Options, EventEmitter], {
      * @param {object} Newman options
      */
     execute: function(requestJSON, options, callback) {
+        
+        log.note("this is the new version");
+        
         // var collectionParseError = Validator.validateJSON('c',requestJSON);
         // if(!collectionParseError.status) {
         //     Errors.terminateWithError("Not a valid POSTMAN collection");
@@ -86,165 +90,12 @@ var Newman = jsface.Class([Options, EventEmitter], {
         // setup the iteration runner with requestJSON passed and options
         this.iterationRunner = new IterationRunner(requestJSON, this.getOptions());
 
- if (options.exportJSON) {
-
-            var changes = false;
-
-            var collectionName = requestJSON.name.replace(/ /g, '');
-            requestJSON.name = collectionName;
-
-            _und.each(requestJSON.folders, function (folder) {
-
-                var folderItemName = folder.name.replace(/ /g, '')
-                if (folderItemName != folder.name) {
-                    changes = true;
-                    folder.name = folderItemName;
-                }
-            });
-
-            if (changes) {
-                console.log('updated fodler names');
-                this.iterationRunner = new IterationRunner(requestJSON, this.getOptions());
-            }
-
-            //Get a ordered list of all the requests
-            var orderedCollection = this.iterationRunner._getOrderedCollection(requestJSON);
-
-
-            _und.each(orderedCollection, function (postmanRequest) {
-                var fileName
-
-                //Remove spaces from Request Name
-                var requestItem = _und.find(requestJSON.requests, function (request) {
-                    return request.id === postmanRequest.id;
-                });
-
-                requestItem.name = postmanRequest.name.replace(/ /g, '')
-                if (requestItem.name != postmanRequest.name) {
-                    changes = true;
-                }
-
-                //Add folder name to path if it exists
-                if (postmanRequest.folderName) {
-                    console.log('Request:%s.%s with ID:%s', postmanRequest.folderName, requestItem.name, postmanRequest.id);
-                    fileName = postmanRequest.folderName + '.' + requestItem.name;
-
-                } else {
-                    console.log('no folder name');
-                    console.log('Request:%s with ID:%s', postmanRequest.name, postmanRequest.id);
-                    fileName = requestItem.name;
-                }
-
-                //build a base file name for data output
-                baseFileName = path.join(options.exportJSON, fileName);
-
-                if (requestItem.tests) {
-                    //build a file name for the test data output
-                    var testsDatafileName = baseFileName + '.tests.json';
-
-                    var testsJSON = requestItem.tests
-                        //write out the info
-                    fs.writeFileSync(testsDatafileName, testsJSON);
-                    console.log('dump tests to file: %s', testsDatafileName);
-                    requestItem.tests = 'file:' + testsDatafileName;
-                }
-
-
-                if (requestItem.rawModeData) {
-                    var rawModeDatafileName = baseFileName + '.rawModeData.json';
-                    var rawModeData = requestItem.rawModeData
-                        //write out the info
-                    fs.writeFileSync(rawModeDatafileName, rawModeData);
-                    console.log('dump rawModeData to file: %s', rawModeDatafileName);
-                    requestItem.rawModeData = 'file:' + rawModeDatafileName;
-                }
-
-
-            }, this);
-
-            if (changes) {
-                var collectionFileName = path.join(options.exportJSON, collectionName + '.colleciton.json');
-                fs.writeFileSync(collectionFileName, JSON.stringify(requestJSON, null, 1));
-                console.log('done something');
-            }
-
+        if (options.exportJSON) {
+            importer.exportJSON(requestJSON, options);
         } 
         else if (options.importJSON) {
-
-            /*take the path supplied and make a single output file which is a postman collection
-            
-               find the file name of colelction it is in the new options.collectionFileName
-            
-               replace the file: refereces with the actual contents of the file for
-                    tests
-                    rawModeData
-                    
-                write a new output file 
-            */
-            
-            console.log('options.collectionFileName:', options.collectionFileName);
-
-            var changes = false;
- 
-            //Get a ordered list of all the requests
-            var orderedCollection = this.iterationRunner._getOrderedCollection(requestJSON);
-
-            _und.each(orderedCollection, function (postmanRequest) {
-
-                var fileName
-
-                //get the single request item
-                var requestItem = _und.find(requestJSON.requests, function (request) {
-                    return request.id === postmanRequest.id;
-                });
-
-                console.log('processing request :%s, tests:%s', requestItem.name,requestItem.tests);
-                
-                    
-
-                //take a file reference and convert to postman
-                if (requestItem.tests && requestItem.tests.startsWith('file:')) {
-                    var fileName = requestItem.tests.substring(5);
-                    console.log('Reading in tests from file:%s', fileName);
-
-                    try {
-                        requestItem.tests = fs.readFileSync(fileName, 'utf8');
-                        changes =true;
-                    } catch (e) {
-                        if (e.code === 'ENOENT') {
-                            console.log('File not found!');
-                        } else {
-                            throw e;
-                        }
-                    }
-                } 
-
-                //take a file reference and convert to postman
-                if (requestItem.rawModeData && requestItem.rawModeData.startsWith('file:')) {
-                    var fileName = requestItem.rawModeData.substring(5);
-                    console.log('Reading in rawModeData from file:%s', fileName);
-
-                    try {
-                        requestItem.rawModeData = fs.readFileSync(fileName, 'utf8');
-                        changes =true;
-                    } catch (e) {
-                        if (e.code === 'ENOENT') {
-                            console.log('File not found!');
-                        } else {
-                            throw e;
-                        }
-                    }
-                } 
-
-            }, this);
-
-            if (changes) {
-                var collectionFileName = options.collectionFileName + '.out';
-                fs.writeFileSync(collectionFileName, JSON.stringify(requestJSON, null, 1));
-                console.log('done something');
-            }
-
-        } else {
+            importer.importJSON(requestJSON, options);
+        }else {
             this.iterationRunner.execute();
         }
